[CL-2440] Move util functions into config
