+++ front/app/components/FormBuilder/components/FormBuilderSettings/index.tsx
@@ -250,6 +250,7 @@ const FormBuilderSettings = ({
             onClose={onClose}
             isDeleteDisabled={isDeleteDisabled}
             onDelete={deleteField}
+            builderConfig={builderConfig}
           />
         )}
         {showTabbedSettings && currentTab === 'logic' && (
