+++ front/app/containers/Admin/projects/project/inputForm/utils.tsx
@@ -1,5 +1,6 @@
 // types
 import { FormBuilderConfig } from 'components/FormBuilder/utils';
+import { IFlatCustomFieldWithIndex } from 'services/formCustomFields';
 
 // intl
 import messages from './messages';
@@ -9,11 +10,38 @@ export const ideationConfig: FormBuilderConfig = {
   viewFormLinkCopy: messages.viewFormLinkCopy,
   formSavedSuccessMessage: messages.successMessage,
   toolboxTitle: messages.addFormContent,
-
   toolboxFieldsToExclude: ['page', 'file_upload'],
   formCustomFields: undefined,
-
   showStatusBadge: false,
   isLogicEnabled: false,
   isEditPermittedAfterSubmissions: true,
+
+  isRequiredToggleAllowed: (field: IFlatCustomFieldWithIndex) => {
+    const keysAlwaysRequired: string[] = ['title_multiloc', 'body_multiloc'];
+    return keysAlwaysRequired.includes(field.key);
+  },
+  isResponseToggleAllowed: (field: IFlatCustomFieldWithIndex) => {
+    const keysResponsesAlwaysShown: string[] = [
+      'title_multiloc',
+      'body_multiloc',
+      'idea_images_attributes',
+    ];
+    return keysResponsesAlwaysShown.includes(field.key);
+  },
+  isEnabledToggleAllowed: (field: IFlatCustomFieldWithIndex) => {
+    const keysAlwaysEnabled: string[] = [
+      'title_multiloc',
+      'body_multiloc',
+      'idea_images_attributes',
+    ];
+    return keysAlwaysEnabled.includes(field.key);
+  },
+  isTitleConfigurable: (field: IFlatCustomFieldWithIndex) => {
+    const keysWithoutConfigurableTitle: string[] = [
+      'title_multiloc',
+      'body_multiloc',
+      'idea_images_attributes',
+    ];
+    return !keysWithoutConfigurableTitle.includes(field.key);
+  },
 };
